sidecar/pyproject.toml
[project]
name = "ovo-sidecar"
version = "0.0.1"
description = "OVO Python sidecar — FastAPI server exposing MLX LLM runtime"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "mlx-lm>=0.20.0",
    "mlx-vlm>=0.4.0",
    "huggingface-hub>=0.26.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sse-starlette>=2.1.3",
    "httpx>=0.27.0",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.0",
    "tomli-w>=1.0.0",
    "rich>=13.9.0",
    "pillow>=10.0.0",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "duckduckgo-search>=6.0.0",
    # [START] Phase 6.4 — local embeddings for semantic Wiki search
    "sentence-transformers>=3.0.0",
    # [END]
    # [START] Phase 8 — Voice I/O: Whisper STT + microphone recording
    "mlx-whisper>=0.4.0",
    "sounddevice>=0.5.0",
    "soundfile>=0.12.0",
    # [END]
    # [START] Phase 7 — text-to-image diffusion runner
    "diffusers>=0.30.0",
    "transformers>=4.44.0",
    "accelerate>=0.33.0",
    "safetensors>=0.4.5",
    # [END]
    # [START] Phase 8.4 — grammar-constrained JSON generation.
    # outlines wraps mlx-lm with a logits processor that masks any token
    # which would break the target JSON schema. The only way local MLX
    # models reliably emit parseable tool calls regardless of weights.
    "outlines>=0.1.0",
    # [END]
    # [START] Phase 5 — PDF text extraction.
    # pdfjs-dist repeatedly chokes inside the Tauri WebView (ReadableStream
    # transport, worker URL resolution), so PDF parsing runs server-side
    # via PyMuPDF (aka fitz). Best-in-class for plain text extraction,
    # handles scanned / multi-column / CJK PDFs that pdfjs fumbles on.
    "pymupdf>=1.24.0",
    # [END]
]

[project.scripts]
ovo-sidecar = "ovo_sidecar.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ovo_sidecar"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "pyinstaller>=6.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
